+++ apps/x/packages/core/src/channels/service.ts
@@ -5,7 +5,7 @@ import type { ChannelsConfig, ChannelsStatus } from "@x/shared/dist/channels.js"
 import container from "../di/container.js";
 import { WorkDir } from "../config/config.js";
 import type { ISessions } from "../sessions/api.js";
-import type { TurnEventHub } from "../turns/event-hub.js";
+import type { ITurnEventBus } from "../turns/event-hub.js";
 import { isSignedIn } from "../account/account.js";
 import { listGatewayModels } from "../models/gateway.js";
 import { listOnboardingModels } from "../models/models-dev.js";
@@ -98,7 +98,7 @@ function ensureBridge(): ChannelBridge {
     if (!bridge) {
         bridge = new ChannelBridge({
             sessions: container.resolve<ISessions>("sessions"),
-            turnEventBus: container.resolve<TurnEventHub>("turnEventBus"),
+            turnEventBus: container.resolve<ITurnEventBus>("turnEventBus"),
             listModels: listBridgeModels,
         });
     }
